Refactor ErrorController to actions.
# Conflicts:
#	module/VuFind/config/module.config.php
#	module/VuFind/src/VuFind/Action/PluginManager.php
