+++ module/VuFind/config/module.config.php
@@ -169,7 +169,6 @@ $config = [
         'factories' => [
             'VuFind\Controller\EITController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\EPFController' => 'VuFind\Controller\AbstractBaseFactory',
-            'VuFind\Controller\ErrorController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\Search2Controller' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\HierarchyController' => 'VuFind\Controller\AbstractBaseFactory',
             'VuFind\Controller\HoldsController' => 'VuFind\Controller\HoldsControllerFactory',
@@ -207,8 +206,6 @@ $config = [
             'eit' => 'VuFind\Controller\EITController',
             'EPF' => 'VuFind\Controller\EPFController',
             'epf' => 'VuFind\Controller\EPFController',
-            'Error' => 'VuFind\Controller\ErrorController',
-            'error' => 'VuFind\Controller\ErrorController',
             'Search2' => 'VuFind\Controller\Search2Controller',
             'search2' => 'VuFind\Controller\Search2Controller',
             'Hierarchy' => 'VuFind\Controller\HierarchyController',
